# docs/swagger.yaml
basePath: /api/v2
definitions:
  internal_handlers.addMemberRequest:
    properties:
      user_id:
        type: integer
    type: object
...
  internal_handlers.adminOIDCAutoProvisionRequest:
    properties:
      auto_provision:
        type: boolean
      default_group_id:
        type: integer
    type: object
...
  internal_handlers.createGroupRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permission_level:
        type: string
    type: object
...
  internal_handlers.groupMemberResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  internal_handlers.groupResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      external_id:
        description: SCIM-provisioned groups carry this
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      member_count:
        type: integer
      name:
        type: string
      permission_level:
        type: string
    type: object
...
  internal_handlers.updateGroupRequest:
    properties:
      description:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      permission_level:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Enterprise gateway for Autonomi decentralized storage. Provides file
    upload/download, user management, API tokens, webhooks, and admin controls.
  title: Indelible API
  version: "2.0"
paths:
  /admin/analytics/costs:
    get:
      description: Return cost analytics broken down by token and department
      parameters:
      - description: Number of days to look back (default 7)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get cost analytics
      tags:
      - 'Admin: Analytics'
  /admin/analytics/tokens:
    get:
      description: Return API token usage statistics for the specified time period
      parameters:
      - description: Number of days to look back (default 7)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get token analytics
      tags:
      - 'Admin: Analytics'
  /admin/analytics/uploads:
    get:
      description: Return upload statistics for the specified time period
      parameters:
      - description: Number of days to look back (default 7)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get upload analytics
      tags:
      - 'Admin: Analytics'
  /admin/groups:
    get:
      description: Return all permission groups with member counts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/internal_handlers.groupResponse'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all groups
      tags:
      - 'Admin: Groups'
    post:
      consumes:
      - application/json
      description: Create a new permission group with a specified permission level
      parameters:
      - description: Group details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.groupResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Group name taken
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a group
      tags:
      - 'Admin: Groups'
  /admin/groups/{id}:
    delete:
      description: Remove a permission group and all its memberships
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a group
      tags:
      - 'Admin: Groups'
    put:
      consumes:
      - application/json
      description: Update a group's name, description, permission level, or active
        status
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated group fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.groupResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a group
      tags:
      - 'Admin: Groups'
  /admin/groups/{id}/members:
    get:
      description: Return the resolved members of a group (id, email, display name).
        V2-304.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/internal_handlers.groupMemberResponse'
              type: array
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List group members
      tags:
      - 'Admin: Groups'
    post:
      consumes:
      - application/json
      description: Add a user to a permission group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: User to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.addMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Already a member
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a group member
      tags:
      - 'Admin: Groups'
  /admin/groups/{id}/members/{userId}:
    delete:
      description: Remove a user from a permission group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID to remove
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not a member
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a group member
      tags:
      - 'Admin: Groups'
  /admin/logs/audit:
    get:
      description: Return audit log entries with optional filtering (permanent, never
        deleted)
      parameters:
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: since
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: until
        type: string
      - description: Max results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Query audit logs
      tags:
      - 'Admin: Logs'
  /admin/logs/config:
    get:
      description: |-
        Return system log entries with optional filtering (retention-managed)
        Return config_audit entries showing old/new value, actor, IP, and UA for every settings change. Permanent (never purged).
      parameters:
      - description: Filter by log level
        in: query
        name: level
        type: string
      - description: Filter by component
        in: query
        name: component
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: since
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: until
        type: string
      - description: Max results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by setting key
        in: query
        name: setting_key
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: since
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: until
        type: string
      - description: Max results (default 100, max 500)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Query config-change audit log
      tags:
      - 'Admin: Logs'
      - 'Admin: Logs'
  /admin/logs/system:
    get:
      responses: {}
      security:
      - BearerAuth: []
  /admin/logs/user:
    get:
      description: Return user activity log entries with optional filtering
      parameters:
      - description: Filter by user ID
        in: query
        name: user_id
        type: integer
      - description: Start date (YYYY-MM-DD)
        in: query
        name: since
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: until
        type: string
      - description: Max results
        in: query
        name: limit
        type: integer
      - description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Query user activity logs
      tags:
      - 'Admin: Logs'
  /admin/oidc/providers:
    get:
      description: Return all configured OIDC identity providers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/internal_handlers.oidcProviderResponse'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List OIDC providers
      tags:
      - 'Admin: OIDC'
    post:
      consumes:
      - application/json
      description: Register a new OIDC identity provider with encrypted client secret
      parameters:
      - description: OIDC provider details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createOIDCRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.oidcProviderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an OIDC provider
      tags:
      - 'Admin: OIDC'
  /admin/oidc/providers/{id}:
    delete:
      description: Remove an OIDC identity provider
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete an OIDC provider
      tags:
      - 'Admin: OIDC'
    put:
      consumes:
      - application/json
      description: Update an existing OIDC provider's configuration (empty client_secret
        keeps existing)
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated provider fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.updateOIDCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.oidcProviderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an OIDC provider
      tags:
      - 'Admin: OIDC'
  /admin/oidc/providers/{id}/auto-provision:
    put:
      consumes:
      - application/json
      description: Configure whether unknown sub/email pairs create new local users
        (off by default), and which group new users join
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      - description: auto_provision + default_group_id (0 to clear)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.adminOIDCAutoProvisionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Toggle OIDC auto-provisioning
      tags:
      - 'Admin: OIDC'
  /admin/oidc/providers/{id}/extra-params:
    put:
      consumes:
      - application/json
      description: Replace the IdP-specific params map appended to the authorize URL.
        Used to set Google Workspace hd=, Microsoft prompt=, AAD domain_hint, etc.
        Send an empty object to clear.
      parameters:
      - description: Provider ID
        in: path
        name: id
        required: true
        type: integer
      - description: extra_authorize_params map (empty {} to clear)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.adminOIDCExtraAuthorizeParamsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set OIDC extra authorize-URL params
      tags:
      - 'Admin: OIDC'
  /admin/quotas:
    get:
      description: Return all configured storage quotas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/internal_handlers.quotaResponse'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List all quotas
      tags:
      - 'Admin: Quotas'
    post:
      consumes:
      - application/json
      description: Create a new storage quota for an entity type
      parameters:
      - description: Quota details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.createQuotaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.quotaResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Duplicate quota
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a quota
      tags:
      - 'Admin: Quotas'
...
  /admin/settings/export:
    get:
      description: Download all settings as a JSON file for backup or migration
      produces:
      - application/json
      responses:
        "200":
          description: JSON settings export
          schema:
            type: file
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Export settings
      tags:
      - 'Admin: Settings'
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT or API token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"